fix: satisfy onboard growth guardrail
